Supervision : Part 2B + System Setting : Part 3
completing the staff management insert, update and delete part with the import data using excel function. For excel purpose, there is a need to update the department setting by adding the department code field for user to easily type the code of the department in excel rather then type the autogenerated id. Next to continue  on making export, custom export for staff management, filter part and the last module for supervision that is supervision arrangement.