Supervision : Part 3B
For this progress, the design of the supervision arrangement module has been successfully completed. Only need works on making the assignment of the supervisors and some validation checks. Update title of research function also completed. Next to focus on this module, making the export supervision, then focus on student management part for import student with supervision data.