Supervision : Part 4A
finalizing the student management module by adding the import and export function, change the design of datatable, the way data to show to user, and customing the export format for both supervision and student list. Next to focus on staff management module ,check all the codes and export format. Moreover, Profile photo settings in student management has been enhance by adding reset button for user to reset the student photo if needed. This will be implemented on staff as well.