Programme Manager : Part 1A
The first step of student portal, setup the route for the programme overview, with a responsive and user freindly design to display the programme activity. Making the path for viewing the flowchart in more secure manner. Adding the submission list page [not completed yet]. Next to focus on submission list page and submission area page before moving to SOP module in staff portal.