Student Semester Assignment : Part 1
Successfully add the initial route, interface design, and the model. The ERD and FRDD have been designed to achieve this module. The semester_id currently as FK in students table must be remove as the relationship change to many-to-many approach for more logical relationship between them. So the Student Semester Assignment page have received the Import function. Just require update on the student semester count. The work still remaining at the semester student list for the semester status update and delete student functionalities.