Submission Approval : Part 3D
Adjusting logic in approval part. If the user were not configured to sign in the modal but still need to sign, their approval claims will be shown in the form not as the signature, but as a text like Approve by Committee/Deputy Dean. several changes in form design as well to cater this issue. Next to focus on the Submission Suggestion.