Submission Suggestion : Part 1C
the multiple approval features has been implemented. The works are still required for this system is the email part need to align with suitable design and wording. For this Submission suggestion module also just the email part need to be completed. Next to focus on the new module that is the nomination part where all the activity that have evaluation requires a nomination process to select the respective chair and examiner. Will implemented soon.